remove dgraph mutate
